pyproject.toml
[tool.poetry]
name = "filamentcolors.xyz"
version = "0.1.0"
description = "The source code for a small website to compare pieces of printed filament."
authors = ["Joe Kaufeld <joe.kaufeld@gmail.com>"]
license = "MIT"

[tool.poetry.dependencies]
python = "^3.7"
django = "^2.1"
Pillow = "^5.3"
django-smartfields = "^1.0"
opencv-python = "^3.4"
scikit-image = "^0.14.1"
numpy = "^1.15"
sorl-thumbnail = "^12.5"
readline = "^6.2"
gunicorn = "^19.9"
psycopg2-binary = "^2.7"
python-dotenv = "^0.10.1"
python-twitter = "^3.5"
colorcorrect = "^0.8.0"
colormath = "^3.0"
djangorestframework = "^3.10"
django-cors-headers = "^3.1"

[tool.poetry.dev-dependencies]
better-exceptions = "^0.2.1"
ipdb = "^0.11.0"
beautifulsoup4 = "^4.7"

[build-system]
requires = ["poetry>=0.12"]
build-backend = "poetry.masonry.api"
